fix 32bit-ism that gave a warning on amd64.

+++ tests/test_input.c
@@ -71,7 +71,7 @@ int main(int argc, char *argv[])
 			case TREECC_TOKEN_IDENTIFIER:
 			{
 				TreeCCDebug(input.linenum, "identifier (len = %d): %s",
-				       	    strlen(input.text), input.text);
+				       	    (int)strlen(input.text), input.text);
 				if(!strcmp(input.text, "parse_literal"))
 				{
 					input.parseLiteral = 1;
